server/requirements.txt
fastapi
uvicorn
pytest
httpx
sqlalchemy
alembic
psycopg2-binary
python-jose[cryptography]
google-auth
google-auth-oauthlib
google-auth-httplib2
python-dotenv
passlib[bcrypt]